BUG: number of atoms along edge in Fourier transform script
The number of atoms along the edge of the configuration
was calculated as block_matrix_shape / block_size, but
this is actually the total number of surface atoms, i.e.
the square of the number of atoms along one edge in the
case of a simple cubic surface lattice. Thus, one needs
to take the square root.